project card legacy to new
