+++ frontend/src/component/project/NewProjectCard/ProjectCardFooter/ProjectCardFooter.tsx
