+++ frontend/src/component/project/NewProjectCard/ProjectCardIcon/ProjectCardIcon.tsx
@@ -38,7 +38,7 @@ export const ProjectCardIcon: VFC<IProjectCardIconProps> = ({ mode }) => {
         return (
             <StyledIconBox data-loading>
                 <HtmlTooltip
-                    title='This project's collaboration mode is set to private. The project and associated feature flags can only be seen by members of the project.'
+                    title="This project's collaboration mode is set to private. The project and associated feature flags can only be seen by members of the project."
                     arrow
                 >
                     <StyledVisibilityIcon />
@@ -51,7 +51,7 @@ export const ProjectCardIcon: VFC<IProjectCardIconProps> = ({ mode }) => {
         return (
             <StyledIconBox data-loading>
                 <HtmlTooltip
-                    title='This project's collaboration mode is set to protected. Only admins and project members can submit change requests.'
+                    title="This project's collaboration mode is set to protected. Only admins and project members can submit change requests."
                     arrow
                 >
                     <StyledLockIcon />
