+++ frontend/src/component/project/ProjectCard/ProjectCard.tsx
@@ -1,14 +1,18 @@
-import type React from 'react';
 import { Menu, MenuItem } from '@mui/material';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import { type SyntheticEvent, useState } from 'react';
+import type React from 'react';
+import { type SyntheticEvent, useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { getProjectEditPath } from 'utils/routePathHelpers';
 import PermissionIconButton from 'component/common/PermissionIconButton/PermissionIconButton';
 import { UPDATE_PROJECT } from 'component/providers/AccessProvider/permissions';
+import AccessContext from 'contexts/AccessContext';
 import { DEFAULT_PROJECT_ID } from 'hooks/api/getters/useDefaultProject/useDefaultProjectId';
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
+import useProjects from 'hooks/api/getters/useProjects/useProjects';
+import { useFavoriteProjectsApi } from 'hooks/api/actions/useFavoriteProjectsApi/useFavoriteProjectsApi';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
+import { FavoriteIconButton } from 'component/common/FavoriteIconButton/FavoriteIconButton';
 import { DeleteProjectDialogue } from '../Project/DeleteProject/DeleteProjectDialogue';
 import {
     StyledProjectCard,
@@ -16,13 +20,14 @@ import {
     StyledBox,
     StyledH2Title,
     StyledEditIcon,
+    StyledProjectIcon,
     StyledDivInfo,
     StyledDivInfoContainer,
     StyledParagraphInfo,
-    StyledProjectCardBody,
-} from './ProjectCard.styles';
-import { ProjectCardFooter } from './ProjectCardFooter/ProjectCardFooter';
-import { ProjectCardIcon } from './ProjectCardIcon/ProjectCardIcon';
+    StyledIconBox,
+} from './LegacyProjectCard.styles';
+import useToast from 'hooks/useToast';
+import { HiddenProjectIconWithTooltip } from '../Project/HiddenProjectIconWithTooltip/HiddenProjectIconWithTooltip';
 
 interface IProjectCardProps {
     name: string;
@@ -45,90 +50,118 @@ export const ProjectCard = ({
     mode,
     isFavorite = false,
 }: IProjectCardProps) => {
+    const { hasAccess } = useContext(AccessContext);
+    const { setToastApiError } = useToast();
     const { isOss } = useUiConfig();
     const [anchorEl, setAnchorEl] = useState<Element | null>(null);
     const [showDelDialog, setShowDelDialog] = useState(false);
     const navigate = useNavigate();
+    const { favorite, unfavorite } = useFavoriteProjectsApi();
+    const { refetch } = useProjects();
 
     const handleClick = (event: React.SyntheticEvent) => {
         event.preventDefault();
         setAnchorEl(event.currentTarget);
     };
 
+    const onFavorite = async (e: React.SyntheticEvent) => {
+        e.preventDefault();
+        try {
+            if (isFavorite) {
+                await unfavorite(id);
+            } else {
+                await favorite(id);
+            }
+            refetch();
+        } catch (error) {
+            setToastApiError('Something went wrong, could not update favorite');
+        }
+    };
+
     return (
         <StyledProjectCard onMouseEnter={onHover}>
-            <StyledProjectCardBody>
-                <StyledDivHeader data-loading>
-                    <ProjectCardIcon mode={mode} />
-                    <StyledBox>
-                        <StyledH2Title>{name}</StyledH2Title>
-                    </StyledBox>
-                    <PermissionIconButton
-                        style={{ transform: 'translateX(7px)' }}
-                        permission={UPDATE_PROJECT}
-                        hidden={isOss()}
-                        projectId={id}
-                        data-loading
-                        onClick={handleClick}
-                        tooltipProps={{
-                            title: 'Options',
-                        }}
-                    >
-                        <MoreVertIcon />
-                    </PermissionIconButton>
-
-                    <Menu
-                        id='project-card-menu'
-                        open={Boolean(anchorEl)}
-                        anchorEl={anchorEl}
-                        style={{ top: 0, left: -100 }}
-                        onClick={(event) => {
-                            event.preventDefault();
-                        }}
-                        onClose={(event: SyntheticEvent) => {
-                            event.preventDefault();
-                            setAnchorEl(null);
-                        }}
-                    >
-                        <MenuItem
-                            onClick={(e) => {
-                                e.preventDefault();
-                                navigate(getProjectEditPath(id));
-                            }}
-                        >
-                            <StyledEditIcon />
-                            Edit project
-                        </MenuItem>
-                    </Menu>
-                </StyledDivHeader>
-                <StyledDivInfo>
-                    <StyledDivInfoContainer>
-                        <StyledParagraphInfo data-loading>
-                            {featureCount}
-                        </StyledParagraphInfo>
-                        <p data-loading>toggles</p>
-                    </StyledDivInfoContainer>
-                    <StyledDivInfoContainer>
-                        <StyledParagraphInfo data-loading>
-                            {health}%
-                        </StyledParagraphInfo>
-                        <p data-loading>health</p>
-                    </StyledDivInfoContainer>
-
-                    <ConditionallyRender
-                        condition={id !== DEFAULT_PROJECT_ID}
-                        show={
-                            <StyledDivInfoContainer>
-                                <StyledParagraphInfo data-loading>
-                                    {memberCount}
-                                </StyledParagraphInfo>
-                                <p data-loading>members</p>
-                            </StyledDivInfoContainer>
-                        }
+            <StyledDivHeader data-loading>
+                <StyledBox>
+                    <FavoriteIconButton
+                        onClick={onFavorite}
+                        isFavorite={isFavorite}
+                        size='medium'
+                        sx={{ ml: -1 }}
                     />
-                </StyledDivInfo>
-            </StyledProjectCardBody>
-            <ProjectCardFooter id={id} isFavorite={isFavorite} />
+                    <StyledH2Title>{name}</StyledH2Title>
+                </StyledBox>
+
+                <PermissionIconButton
+                    style={{ transform: 'translateX(7px)' }}
+                    permission={UPDATE_PROJECT}
+                    hidden={isOss()}
+                    projectId={id}
+                    data-loading
+                    onClick={handleClick}
+                    tooltipProps={{
+                        title: 'Options',
+                    }}
+                >
+                    <MoreVertIcon />
+                </PermissionIconButton>
+
+                <Menu
+                    id='project-card-menu'
+                    open={Boolean(anchorEl)}
+                    anchorEl={anchorEl}
+                    style={{ top: 0, left: -100 }}
+                    onClick={(event) => {
+                        event.preventDefault();
+                    }}
+                    onClose={(event: SyntheticEvent) => {
+                        event.preventDefault();
+                        setAnchorEl(null);
+                    }}
+                >
+                    <MenuItem
+                        onClick={(e) => {
+                            e.preventDefault();
+                            navigate(getProjectEditPath(id));
+                        }}
+                    >
+                        <StyledEditIcon />
+                        Edit project
+                    </MenuItem>
+                </Menu>
+            </StyledDivHeader>
+            <StyledIconBox data-loading>
+                <ConditionallyRender
+                    condition={mode === 'private'}
+                    show={<HiddenProjectIconWithTooltip />}
+                    elseShow={<StyledProjectIcon />}
+                />
+            </StyledIconBox>
+            <StyledDivInfo>
+                <StyledDivInfoContainer>
+                    <StyledParagraphInfo data-loading>
+                        {featureCount}
+                    </StyledParagraphInfo>
+                    <p data-loading>toggles</p>
+                </StyledDivInfoContainer>
+                <StyledDivInfoContainer>
+                    <StyledParagraphInfo data-loading>
+                        {health}%
+                    </StyledParagraphInfo>
+                    <p data-loading>health</p>
+                </StyledDivInfoContainer>
+
+                <ConditionallyRender
+                    condition={id !== DEFAULT_PROJECT_ID}
+                    show={
+                        <StyledDivInfoContainer>
+                            <StyledParagraphInfo data-loading>
+                                {memberCount}
+                            </StyledParagraphInfo>
+                            <p data-loading>members</p>
+                        </StyledDivInfoContainer>
+                    }
+                />
+            </StyledDivInfo>
             <DeleteProjectDialogue
                 project={id}
                 open={showDelDialog}
